fix(cleanup): Cast stat.Dev to uint64 for Darwin compatibility
Darwin syscall.Stat_t.Dev is int32 while Linux is uint64.
Explicit cast fixes cross-platform build.

+++ plugins/fs.go
@@ -16,7 +16,7 @@ func deviceID(path string) (uint64, error) {
 	if err := syscall.Stat(path, &stat); err != nil {
 		return 0, err
 	}
-	return stat.Dev, nil
+	return uint64(stat.Dev), nil
 }
 
 // getDirSizeSameDevice calculates directory size without crossing mount boundaries.
